requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.15.4
fsspec==2024.6.1
huggingface-hub==0.24.5
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
keybert==0.8.5
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.4.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
openai-whisper==20231117
packaging==24.1
pillow==11.0.0
pydub==0.25.1
Pygments==2.18.0
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.9.4
safetensors==0.4.4
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers==3.3.1
six==1.16.0
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.66.5
transformers==4.44.0
typing_extensions==4.12.2
urllib3==2.2.2
